revno: 13504 message:   Apply context menu for custom registration form function; Remove registration-date option in reportDateToUse property of Program-stage; minor fix for tabular report.
committer: Tran Chau <tran.hispvietnam@gmail.com>
branch nick: dhis2
timestamp: Tue 2013-12-31 10:05:28 +0700

+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-patient/src/main/java/org/hisp/dhis/patient/action/dataentryform/DelRegistrationFormAction.java
@@ -80,7 +80,7 @@ public class DelRegistrationFormAction
         throws Exception
     {
         PatientRegistrationForm registrationForm = null;
-        
+
         if ( id != null )
         {
             Program program = programService.getProgram( id );
@@ -91,9 +91,12 @@ public class DelRegistrationFormAction
         {
             registrationForm = patientRegistrationFormService.getCommonPatientRegistrationForm();
         }
-        
-        patientRegistrationFormService.deletePatientRegistrationForm( registrationForm );
 
+        if ( registrationForm != null )
+        {
+            patientRegistrationFormService.deletePatientRegistrationForm( registrationForm );
+        }
+        
         return SUCCESS;
     }
 }
